requirements.txt
flask>=2.0.0
docxtpl>=0.16.0
pandas>=1.3.0
openpyxl>=3.0.0
waitress>=2.1.0
